requirements.txt
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
dialogflow==1.1.0
google-api-core==1.23.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
urllib3==1.26.1
